requirements.txt
colorama==0.4.6
delta-spark==3.3.0
importlib_metadata==8.6.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
py4j==0.10.9.7
pyspark==3.5.4
pytest==8.3.4
zipp==3.21.0
